Extracting Queriable and Solvable concerns
