+++ app/models/concerns/submittable/queriable.rb
@@ -1,7 +1,5 @@
-module WithQueries
-
+module Queriable
   def submit_query!(params)
     Query.new(query: params[:query], content: params[:content], exercise: self)
   end
-
 end
