+++ app/models/concerns/with_assignments.rb
@@ -33,14 +33,6 @@ module WithAssignments
     assignment_for(user).try(&:submissions_count) || 0
   end
 
-  def submit_solution(user, solution_attributes={})
-    solution = Solution.new(solution_attributes)
-    assignment = transaction do
-      assignment_with(user, solution: solution.content).tap { |it| it.submit! }
-    end
-    solution.assignment = assignment
-    solution
-  end
 
   private
