+++ app/models/exercise.rb
@@ -14,7 +14,8 @@ class Exercise < ActiveRecord::Base
           WithAssignments, WithGuide,
           WithLocale, WithExpectations,
           WithLanguage,
-          WithQueries
+          Queriable,
+          Solvable
   extend FriendlyId
 
   friendly_id :generate_custom_slug, use: [:slugged, :finders]
